requirements.txt
#pipreqs core requirements
click
dvc
dvc[gdrive]
dvc[gs]
omegaconf
python-dotenv
PyYAML
timm
torch
torchvision
tqdm
wandb

